requirements.txt
matplotlib==3.6.2
ruamel.base==1.0.0
seaborn==0.12.2
sentence_transformers==2.2.2
torch==1.12.1
tqdm==4.64.1
transformers==4.27.2
